Base de datos

+++ Proyecto1/Database/script.sql
@@ -4,12 +4,12 @@ USE proyecto1;
 
 CREATE TABLE IF NOT EXISTS proyecto1.RAM (
     id INT PRIMARY KEY AUTO_INCREMENT,
-    libre INT NOT NULL,
-    ocupada INT NOT NULL
+    libre FLOAT NOT NULL,
+    ocupada FLOAT NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS proyecto1.CPU (
     id INT PRIMARY KEY AUTO_INCREMENT,
-    libre INT NOT NULL,
-    ocupada INT NOT NULL
+    libre FLOAT NOT NULL,
+    ocupada FLOAT NOT NULL
 );
